+++ app/src/main/java/com/example/galgeleg/Model/Spillogik/Galgelogik.java
@@ -1,4 +1,4 @@
-package com.example.galgeleg;
+package com.example.galgeleg.Model.Spillogik;
 
 
 import java.io.BufferedReader;
