+++ app/src/main/java/com/example/galgeleg/View/Fragment2_config.java
@@ -1,6 +1,5 @@
-package com.example.galgeleg;
+package com.example.galgeleg.View;
 
-import android.content.Intent;
 import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -12,7 +11,8 @@ import android.widget.TextView;
 import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
 
-import com.example.galgeleg.Database.Database;
+import com.example.galgeleg.Model.Database.Database;
+import com.example.galgeleg.R;
 
 public class Fragment2_config extends Fragment implements View.OnClickListener {
