+++ app/src/main/java/com/example/galgeleg/View/Fragment_settings.java
@@ -1,4 +1,4 @@
-package com.example.galgeleg;
+package com.example.galgeleg.View;
 
 public class Fragment_settings {
 }
