+++ app/src/main/java/com/example/galgeleg/View/Game/ImageControl.java
@@ -1,7 +1,9 @@
-package com.example.galgeleg;
+package com.example.galgeleg.View.Game;
 
 import android.widget.FrameLayout;
 import android.widget.ImageView;
+
+import com.example.galgeleg.R;
 
 /*
 Image Controle is used to controle the images that are shown in the Fragment3_Game.
@@ -10,8 +12,7 @@ Image Controle is used to controle the images that are shown in the Fragment3_Ga
 public class ImageControl {
 
 
-    ImageControl() {
-    }
+    ImageControl() { }
 
     public void map(int antalforkerte, ImageView errorIm, FrameLayout errorImage) {
